ISSUE-1089 Rename SaaS sender matchers
Better prefix the sender related matchers with "Sender".
