+++ docs/modules/ROOT/pages/tmail-backend/configure/twake-workplace.adoc
@@ -66,13 +66,13 @@ If you use the `PostgresSaaSModule`, you need to add SaaS related columns (`saas
 ALTER TABLE tmail_schema.users ADD COLUMN saas_plan VARCHAR;
 ```
 
-=== HasAnySaaSPlan matcher
+=== SenderHasAnySaaSPlan matcher
 
-The `HasAnySaaSPlan` matcher can be used to determine if the sender of an email has a SaaS plan.
+The `SenderHasAnySaaSPlan` matcher can be used to determine if the sender of an email has a SaaS plan.
 
-=== HasSaaSPlan matcher
+=== SenderHasSaaSPlan matcher
 
-The `HasSaaSPlan` matcher allows to match emails where the sender has a specific SaaS plan.
+The `SenderHasSaaSPlan` matcher allows to match emails where the sender has a specific SaaS plan.
 This is useful for applying mail processing rules based on the sender's SaaS subscription.
 
 To use the matcher, specify a comma-separated list of SaaS plan names as the condition.
@@ -80,7 +80,7 @@ For example, to match senders with either the `standard` or `premium` plan:
 
 [source,xml]
 ----
-<mailet match="HasSaaSPlan=standard,premium" class="..." />
+<mailet match="SenderHasSaaSPlan=standard,premium" class="..." />
 ----
 
 If the sender's SaaS plan matches any of the specified plans, the matcher will return the recipients of the email.
